+++ src/main/java/lotto/domain/LottoPlay.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import lotto.dto.IssueNumber;
 import lotto.dto.LottoNumber;
+import lotto.dto.LottoNumbers;
 
 import java.util.*;
