+++ src/main/java/lotto/dto/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import lotto.dto.LottoNumber;
